requirements.txt
telethon
pydantic
pydantic-settings
PyYAML
pytz
python-dotenv
